apps/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "canon-vault-api"
version = "0.2.0"
description = "Canon Vault API — single source of truth for anime/IP worldbuilding"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "London / Astral Lab"}]
dependencies = [
    "fastapi>=0.115.0",
    "sqlmodel>=0.0.22",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.5.0",
    "rapidfuzz>=3.5.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
